\name{F4}
\docType{data}
\alias{F4}
\title{Genotypes of one million simulated F4 lines.}

\description{These are three-marker genotypes of one million F4 lines
from a cross between parent A and B, simulated using \R package
\pkg{hypred}.}

\usage{F4}

\format{A character matrix with one million rows and three columns.
Homozygousity for parents A or B is coded as "A" and "B",
respectively. Heterozygousity as "H".}

\references{
  Frank Technow(2013). hypred: Simulation of Genomic Data in Applied
  Genetics. R package version 0.4.
}
\keyword{datasets}