\name{map}
\docType{data}
\alias{map}
\title{Genetic map of the three markers in the F4 data set.}

\description{Genetic map of the three markers in the F4 data set. The
  unit is Morgan. This map can be used to compute the recombination
  frequencies between the markers using the inverse of the Haldane
  mapping function.}

\usage{map}

\format{A numeric vector with three elements (\code{c(0.00, 0.05, 0.20)}).}

\references{
  Frank Technow(2013). hypred: Simulation of Genomic Data in Applied
  Genetics. R package version 0.4.
}
\keyword{datasets}